自动新建合同接口；ueserId移动
