+++ server/src/main/java/com/example/stc/service/ContractService.java
@@ -59,7 +59,7 @@ public interface ContractService extends BaseService<Contract>{
     /**
      * 自动创建单个合同
      */
-    Contract newContractAuto(String pid, String uid);
+    Contract newContract(String pid, String uid);
 
     /**
      * 通过合同ID修改单个合同记录
